hr_module 2/hr_module/hr_module/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
asgiref==3.4.1
autopep8==1.6.0
bcrypt==3.2.0
cffi==1.15.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
Unipath==1.1
whitenoise==5.3.0
django-filter==2.4.0
cx_oracle==8.3.0